Fix application crash on invalid tx_id
This was being caused by an invalid uuid being passed in.  UUID has a precise specification
in RFC 4122, this validation makes it follow one of the valid forms of UUID.  The validation
could be expanded in future if we have UUIDs in different forms, but currently we only
use the one.