Update tx_id validation to use a more standard way
The way the tx_id is now validated is identical to the UUID() rule that
is in wtforms.  The reason we're not using it is because it threw an error on the empty
string, and adding optional() for some reason didn't validate it when the string wasn't empty
(leading to crashes again).

The fields in the StoreForm were added so that on the search results being returned, it now
doesn't clear down the tx_id, ru_ref and survey_id fields (making for a slightly better user
experience).